add build files

+++ AOJ/CGL_7/A.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 
-#include "src/point.hpp"
-#include "src/circle.hpp"
-#include "src/intersection.hpp"
+#include "point.hpp"
+#include "circle.hpp"
+#include "intersection.hpp"
 
 using namespace std;
 using namespace sapphre15::geometry;
@@ -11,5 +11,5 @@ int main() {
     Point c1, c2;
     Real r1, r2;
     cin >> c1 >> r1 >> c2 >> r2;
-    cout << intersection(Circle(c1, r1), Circle(c2, r2)) << endl;
+    cout << static_cast<int>(intersection(Circle(c1, r1), Circle(c2, r2))) << endl;
 }
